myapp/requirements.txt
#--find-links https://download.pytorch.org/whl/torch_stable.html
#torch==1.7.0+cpu
#transformers==4.1.1

gensim==3.8.3
newspaper3k==0.2.8
validators==0.18.2
textstat==0.7.0
textblob==0.15.3
Flask==1.1.2
gunicorn==20.0.4
pytest==5.2.1




